INAT-Solutions/requirements.txt
PyQt5
psycopg2-binary
reportlab
Pillow
packaging
svglib
fpdf
beautifulsoup4
pydantic
iso4217
pyban
qrcode
lxml
requests
pandas
msal
paramiko
qrbill
PyMuPDF
